Add clang-format and clang-tidy checks for 2015/2021 C++ projects (#74)
* Add clang-format and clang-tidy checks for 2015/2021 C++ projects

- Add .clang-format (Microsoft-based style) and reformat all tracked
  C++ sources in 2015 and 2021.
- Add .clang-tidy config (bugprone-*, clang-analyzer-*, performance-*)
  with HeaderFilterRegex scoped to project sources so vendored/
  third-party headers (e.g. nlohmann/json fetched via FetchContent)
  are not flagged.
- Fix all resulting clang-tidy findings across both projects,
  including two real bugs in 2021 day04 (use-after-move relying on
  unspecified state, and an ineffective std::move on a const
  reference that silently degraded to a copy).
- Add CI steps (Linux only) to check formatting and run clang-tidy
  as blocking checks in .github/workflows/2015.yml and 2021.yml.

Copilot-Session: 97248fff-8b13-437c-b3a2-d94ca0e9acc5

* Address PR review feedback

- 2021/day15: fix Dijkstra to track cumulative distance-to-node
  instead of the last edge's risk_level, and skip stale queue
  entries. The previous logic could accept a worse path or reject a
  better one for some inputs.
- 2021/day10: guard against calling stack.back() on an empty stack
  when a closing token appears with no matching opener, which was
  undefined behavior.
- CI: only pass .cc files to clang-tidy (headers don't have their own
  compile_commands.json entries and are still checked transitively
  via HeaderFilterRegex), and use xargs --no-run-if-empty so the
  steps don't invoke the tools with an empty file list.

Copilot-Session: 97248fff-8b13-437c-b3a2-d94ca0e9acc5

* Address additional PR review feedback

- 2021/day04: pass chosen_numbers by const reference in score_board
  instead of by value, avoiding an unnecessary copy on every call.
- 2015/2021 .clang-tidy: fix HeaderFilterRegex to allow an arbitrary
  prefix before (day../libcommon), since clang-tidy reports header
  paths as absolute rather than relative to the project root. Without
  this, project header diagnostics could be silently excluded instead
  of just the intended vendored/third-party headers.

Copilot-Session: 97248fff-8b13-437c-b3a2-d94ca0e9acc5

* Fix unsafe <cctype> usage flagged by Copilot review

- 2015/day07/lib.cc: include <cctype> explicitly instead of relying
  on a transitive include, and cast to unsigned char before calling
  std::isdigit (passing a signed char is undefined behavior for
  non-ASCII byte values).
- 2021/day12/lib.cc: include <cctype> and cast to unsigned char
  before calling std::isupper for the same reason.

Copilot-Session: 97248fff-8b13-437c-b3a2-d94ca0e9acc5

---------

Copilot-Session: 97248fff-8b13-437c-b3a2-d94ca0e9acc5

+++ 2015/day03/lib.cc
@@ -45,5 +45,5 @@ int64_t day03::count_visited_houses(const std::string &input, size_t actor_count
         step_count++;
     }
 
-    return visited.size();
+    return static_cast<int64_t>(visited.size());
 }
